Move SignalHandler to the screen module
The SignalHandler class is used only inside of UIScreen.
